depth first ai is working
the depth first ai works. for some reason it starts breaking
when the depth of search goes too high. It might be that it
finds multiple different ways to guarantee a win and just
happens to choose the one that is not the quickest. Could
be fixed by adding a "number of turns" value to each choice
but i don't really know of a super simple way to do that.