bug fix: stops ai from continuously sending cmds
This fixes the bug where the ai would send out tons of commands for each
move. Just needed to make an option for the move function to not
actually send out the command. This way, game does not send out commands
to the robots when it is just trying to get to a new state.